Apps: Address review findings in streaming manifest parser
Six correctness fixes and three defensive-hardening tweaks raised by code review of the preceding commits.

ManifestEnumFlagNames rewritten: the table now references android.R.attr.* directly (eliminating four wrong hand-typed resource IDs — foregroundServiceType / importantForAutofill / gwpAsanMode / usesPermissionFlags were pointing at unrelated framework attrs). protectionLevel is now an EnumWithFlags decoder — the base value (normal/dangerous/signature/signatureOrSystem/internal) lives in the low nibble with flag bits above; the previous pure-flag decoder would render "signatureOrSystem" (data=3) as "dangerous|signature". windowSoftInputMode uses a new TwoNibbleEnum decoder that correctly decodes the state enum in bits 0..3 and the adjust enum in bits 4..7 — previously "stateAlwaysHidden" (data=0x3) rendered as "stateUnchanged|stateHidden". usesPermissionFlags.neverForLocation fixed to the 0x00010000 bit. Tests cover all 12 table entries and assert every key matches its android.R.attr constant.

ApkManifestReader catches IOException / ZipException during the manifest byte read and classifies them as MALFORMED_APK. Corrupt DEFLATE streams in user-installed APKs no longer escape the reader and crash the hint scanner mid-iteration.

ManifestHintRepo policy refined: LOW_MEMORY outcomes preserve stale hints (transient — next scan refreshes), and Failure outcomes now delete stale-version hints (the current-version shortcut at loop top means Failure branch only sees stale entries; showing old-version flags as current is worse than "no data"). Matches the stable-vs-transient split already used by the outcome memory-cache.

ManifestRepo single-flight atomicity: the piggyback check for an in-flight viewer parse and the queries in-flight insertion now live inside one mutex block via a small InFlight sealed class. A viewer parse registering between the two previous lock acquisitions would have caused a duplicate parse.

BinaryXmlStringPool UTF-16 guard: byteLen computed as Long to prevent Int overflow when charLen is near Int.MAX_VALUE (from the two-word varlen encoding). Previously the overflow bypassed the bounds check and allocated a multi-gigabyte CharArray before the caller classified the OOM.

QueriesExtractor resets intentDepth when queriesDepth tears down, so a malformed <queries><intent>...</queries> (unclosed intent) can't leak intent state into unrelated elements after the queries block.

+++ app/src/main/java/eu/darken/myperm/apps/core/manifest/binaryxml/BinaryXmlStringPool.kt
@@ -74,8 +74,11 @@ internal object BinaryXmlStringPool {
         val charLen = readUtf16VarLen(chunk, at, chunkLimit)
         val p = at + if (((readU16(chunk, at)) and 0x8000) != 0) 4 else 2
         if (charLen < 0) throw BinaryXmlException("negative utf16 char length")
-        val byteLen = charLen * 2
-        if (p + byteLen > chunkLimit) throw BinaryXmlException("utf16 string overruns chunk")
+        // Compute byteLen in Long to prevent Int overflow (charLen up to 0x7FFFFFFF from the
+        // two-word varlen encoding; charLen * 2 as Int would wrap to negative, slipping past the
+        // bounds check and allocating a multi-GB CharArray before OOM classifies it).
+        val byteLen = charLen.toLong() * 2L
+        if (p.toLong() + byteLen > chunkLimit.toLong()) throw BinaryXmlException("utf16 string overruns chunk")
         val chars = CharArray(charLen)
         var cp = p
         for (i in 0 until charLen) {
